requirements.txt
pandas>=2.0.0
numpy>=1.24.0,<2.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
matplotlib>=3.7.0
seaborn>=0.12.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
